chore: change imports on simpler libraries
