+++ libs/apollo/logging/src/lib/apolloWinstonLoggingPlugin/index.ts
@@ -1,0 +1,1 @@
+export { apolloWinstonLoggingPlugin } from './apolloWinstonLoggingPlugin'
