+++ libs/nest/decorators/src/lib/CurrentUserId/index.ts
@@ -1,0 +1,1 @@
+export { CurrentUserId } from './CurrentUserId'
