+++ libs/nest/decorators/src/lib/IdAsKey/index.ts
@@ -1,0 +1,1 @@
+export { IdAsKey, idAsKey } from './IdAsKey'
