+++ libs/nest/decorators/src/lib/KeyAsId/index.ts
@@ -1,0 +1,1 @@
+export { KeyAsId, keyAsId } from './KeyAsId'
