+++ libs/nest/decorators/src/lib/Omit/Omit.ts
@@ -1,0 +1,17 @@
+import { omit } from 'lodash'
+
+export function Omit(omitFields: string[]) {
+  return (
+    target: unknown,
+    propertyKey: string,
+    descriptor: PropertyDescriptor
+  ) => {
+    const childFunction = descriptor.value
+    descriptor.value = async function (...args: unknown[]) {
+      const newArgs = Array.isArray(args)
+        ? args.map((r) => omit(r, omitFields))
+        : omit(args, omitFields)
+      return childFunction.apply(this, newArgs)
+    }
+  }
+}
