+++ libs/nest/decorators/src/lib/Omit/index.ts
@@ -1,0 +1,1 @@
+export { Omit } from './Omit'
