+++ libs/nest/gqlAuthGuard/src/lib/GqlAuthGuard/index.ts
@@ -1,0 +1,1 @@
+export { GqlAuthGuard } from './GqlAuthGuard'
