--- libs/shared/util-graphql/src/index.ts
@@ -1,1 +1,0 @@
-export { schemaBuilder } from './lib/schemaBuilder'
